proper model halfing

+++ reinforce/policies/ppo_pixels_cnn.py
@@ -76,7 +76,13 @@ class PPOPixelsCNN(nn.Module):
             self.gru.bias_ih.data.fill_(0)
             self.gru.bias_hh.data.fill_(0)
 
+        self.half()
+
     def forward(self, inputs, hiddens, masks):
+        inputs = inputs.half()
+        hiddens = hiddens.half()
+        masks = masks.half()
+
         x = F.elu(self.cv1(inputs))
         x = F.elu(self.cv2(x))
         # x = F.elu(self.cv3(x))
@@ -111,7 +117,7 @@ class PPOPixelsCNN(nn.Module):
         v = F.tanh(self.fc1_v(x))
         v = self.fc2_v(v)
 
-        return v, a, hiddens
+        return v.float(), a.float(), hiddens.float()
 
     def input_shape(self):
         return (
@@ -131,7 +137,7 @@ class PPOPixelsCNN(nn.Module):
             ),
             axis=-1,
         )
-        observation = torch.from_numpy(observation).half().unsqueeze(1)
+        observation = torch.from_numpy(observation).unsqueeze(1)
 
         return observation
 
@@ -158,7 +164,7 @@ class PPOPixelsCNN(nn.Module):
             if self.fixed_action_std > 0.0:
                 action_std = (
                     self.fixed_action_std *
-                    torch.ones(action_mean.size()).half()
+                    torch.ones(action_mean.size())
                 ).to(self.device)
                 action_logstd = action_std.log()
 
@@ -201,7 +207,7 @@ class PPOPixelsCNN(nn.Module):
             if self.fixed_action_std > 0.0:
                 action_std = (
                     self.fixed_action_std *
-                    torch.ones(action_mean.size()).half()
+                    torch.ones(action_mean.size())
                 ).to(self.device)
                 action_logstd = action_std.log()
